Market News/Market_News_Yahoo_Extract_Function/requirements.txt
pyarrow
yfinance
pandas
google-cloud-bigquery
textblob
flask
